update message protocol

+++ scripts/mysql/lark_user/sqls/20221221123008_create_table_user_locations.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS `user_locations`;
 CREATE TABLE `user_locations` (
   `uid` bigint unsigned NOT NULL COMMENT '用户ID',
-  `longitude` double(10,6) NOT NULL DEFAULT '0.000000' COMMENT '经度',
-  `latitude` double(10,6) NOT NULL DEFAULT '0.000000' COMMENT '纬度',
+  `longitude` DECIMAL(10,6) NOT NULL DEFAULT '0.000000' COMMENT '经度',
+  `latitude` DECIMAL(10,6) NOT NULL DEFAULT '0.000000' COMMENT '纬度',
   `online_ts` bigint NOT NULL DEFAULT '0' COMMENT '最后一次上线时间',
   `created_ts` bigint NOT NULL DEFAULT '0',
   `updated_ts` bigint NOT NULL DEFAULT '0',
